+++ package.json
@@ -29,8 +29,8 @@
     "@polkadot/extension-dapp": "^0.40.4",
     "@polkadot/keyring": "^9.4.1",
     "@polkadot/ui-keyring": "^2.4.1",
-    "@zeitgeistpm/avatara-nft-sdk": "^1.0.0",
-    "@zeitgeistpm/avatara-react": "^1.0.0",
+    "@zeitgeistpm/avatara-nft-sdk": "^1.1.5",
+    "@zeitgeistpm/avatara-react": "^1.1.5",
     "@zeitgeistpm/avatara-util": "^1.0.0-alpha.0",
     "@zeitgeistpm/sdk": "^0.9.1",
     "@zeitgeistpm/sdk-next": "npm:@zeitgeistpm/sdk@2.0.41",
