+++ pages/liquidity/index.tsx
@@ -1,12 +1,16 @@
 import { Sdkv2Context } from "components/context/Sdkv2Context";
 import { observer } from "mobx-react";
 import { NextPage } from "next";
-import { useContext } from "react";
+import { useContext, useEffect } from "react";
 
 const LiquidityPools: NextPage = observer(() => {
   const sdk$ = useContext(Sdkv2Context);
 
-  console.log(sdk$);
+  useEffect(() => {
+    sdk$.subscribe((sdk) => {
+      console.log(sdk);
+    });
+  }, [sdk$]);
 
   return <div>pools</div>;
 });
